Add csv-writer dependency
The csv-writer package has been added as a new dependency. This
will enable writing data to CSV files in a more streamlined and
efficient manner compared to manually constructing CSV strings.

The addition of this dependency suggests upcoming changes or new
features involving the generation and export of data in CSV format.